added amorphous_lld argument to auto.fpf
This was added so that I can easily tune the function for cases where I think organic matter is below the detection limit. I have also added an RMSE function because I keep forgetting what it is. Lastly, I've added the NSIS data frame to the package data so that I can load it easily.